Add code for demonstration of buddy_check titanlib function
Code added for the demonstration of the buddy_check and
buddy_event_check functions in the titanlib package.

Increased the functionality of the plot_wow_data
function by allowing the buffer_val command argument to
take a list in order to only plot a buffer around some
of the stations.

Edited range_check.cpp for the titanlibcustom package to
return the estimated mean temperature values for each
station for a given date/time and coordinates. This should
allow us to check if these values are reasonable and helps
to explicitly demonstrate how range_check_climatology works.