Add functionality for isolating and plotting additional rainfall data
It is now possible to isolate and plot Rainfall Accumulation, Rainfall
Accumulation Hourly and Rainfall Rate. Note that at the time of this
update we still require confirmation on the official Rainfall Rate
Data so this functionality is not working for the official data.

Bux fix that prevented the elevation data being added to the rainfall
geodataframes.