Add python script to run end to end QC checks
The python script WOW_stations_QC.py was created to perform the chosed QC
checks back to back and output the result. This is a draft version of the
final solution.

The Data Importation and Function Definition file has been split into two
different scripts, one with the original name the other called
Live_Station_Data_Importation.py. This is to hopefully streamline the process
for future use, the Data Importation and function Definition file will import
the static data and define functions, the live importation file will need to
be changed by Met Eireann to import the newest station data and eventually
should be rewritten to pull straight from the databases rather than from
csv files.

The other example files have been updated to remove unnecessary and confusing
references to titanlibcustom. Now the only file that uses the custom package
is the Custom Range Check Climatology Examples notebook where it demonstrates
the calculation of the mean climate temperature based on Latitude and
compares the results to the Irish specific climate data.