Add additional plotting functionality to WOW_stations_QC
Added the ability to plot buffers (radii) around the stations that failed
certain checks. So, for stations that fail the fgt or sct res checks
due to an isolated inner circle observation, the inner_radius will be
plotted around them. For stations that fail the fgt or sct res checks
due to an isolated outer circle observations, the outer_radius will be
plotted around them. For the stations that fail the isolation check,
the radius will be plotted around them.